backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
google-play-scraper==1.2.4
openai==1.14.3
fpdf2==2.7.8
chromadb==0.4.24 
pytest==8.1.1  
httpx==0.27.0  
